Add 'category' widget shared value.
Add a "add_subitem_text_template" to the sequence widget for customizing
the add link.

Left-float the close button of sequence items.

Left-float the existing filename of a file upload widget.

Change some CSS ids and classes.

Don't show the title of a sequence item in the sequence widget.

Don't include the <li> of a hidden structure widget.