Add support for html5 native time-related widgets
  * for the DateInputWidget, the .datepicker interfered with
    the native widget shown on field enter

  * the DateTimeInputWidget's template defined the input field
    as being of type "text" instead of "datetime"; therefore,
    the native datitime widget was never shown

To enable the correct behaviour, the jqueryui datepicker will be
activated only after detecting the native support for the respective
input type is missing via a custom generated modernizr.js.

The installed version of modernizr.js was generated enabling only detection
of input types and attributes; this way, the filesize is just 2394 bytes,
instead of 15483 of the full minified distribution.