A better crack at a nicer looking FileUploadWidget.
Very clever ideas blantantly ripped-off from
http://www.abeautifulsite.net/blog/2013/08/whipping-file-inputs-into-shape-with-bootstrap-3/

Issues:

I've not tested this with many browsers.

I'm not sure where to put the id="${oid}".  (Is it really even necessary?)
Putting on the file input element causes deform.js to focus it (if it's
the first field on the page) which, I guess because it is not visible,
results in unwanted scrolling.