deform.js:focusOnFirst() handles 'on', 'off' forms
If a form's data-deform-focus-form attribute is 'off', no focusing
will be done on that form.

If a form's data-deform-focus-form attribute is 'on' AND a child
input has its data-deform-manual-focus attribute set to 'on', that
input will be focused.

If a form's data-deform-focus-form attribute is 'on' AND no child has
its data-deform-manual-focus attribute set to 'on', the first input
of the form will be focused.